# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is ZAManager api list.",
        "title": "ZAManager API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:80",
    "basePath": "/api/v1",
    "paths": {
        "/access/client": {
            "get": {
                "description": "获取ZTA的client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "GetClient",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "put": {
                "description": "修改ZTA的client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "EditClient",
                "parameters": [
                    {
                        "description": "修改ZTA的client",
                        "name": "Client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.EditClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "新增ZTA的client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "AddClient",
                "parameters": [
                    {
                        "description": "新增ZTA的client",
                        "name": "Client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.AddClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/access/client/{id}": {
            "delete": {
                "description": "删除ZTA的client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "DelClient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/access/relay": {
            "get": {
                "description": "获取ZTA的relay",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "GetRelay",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "put": {
                "description": "修改ZTA的relay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "EditRelay",
                "parameters": [
                    {
                        "description": "修改ZTA的relay",
                        "name": "Relay",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.EditRelay"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "新增ZTA的relay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "AddRelay",
                "parameters": [
                    {
                        "description": "新增ZTA的relay",
                        "name": "Relay",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.AddRelay"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/access/relay/{id}": {
            "delete": {
                "description": "删除ZTA的relay",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "DelRelay",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/access/resource": {
            "get": {
                "description": "获取ZTA的resource",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "GetResource",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "put": {
                "description": "修改ZTA的resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "EditResource",
                "parameters": [
                    {
                        "description": "修改ZTA的resource",
                        "name": "Resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.EditResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "新增ZTA的resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "AddResource",
                "parameters": [
                    {
                        "description": "新增ZTA的resource",
                        "name": "Resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.AddResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/access/resource/{id}": {
            "delete": {
                "description": "删除ZTA的resource",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "DelResource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/access/server": {
            "get": {
                "description": "获取ZTA的server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "GetServer",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "put": {
                "description": "修改ZTA的server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "EditServer",
                "parameters": [
                    {
                        "description": "修改ZTA的server",
                        "name": "Server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.EditServer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "新增ZTA的server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "AddServer",
                "parameters": [
                    {
                        "description": "新增ZTA的server",
                        "name": "Server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mparam.AddServer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/access/server/{id}": {
            "delete": {
                "description": "删除ZTA的server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZTA"
                ],
                "summary": "DelServer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Res": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "mmysql.ClientTarget": {
            "type": "object",
            "required": [
                "host",
                "port"
            ],
            "properties": {
                "host": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
        "mparam.AddClient": {
            "type": "object",
            "required": [
                "expire",
                "name",
                "port",
                "server_id",
                "target"
            ],
            "properties": {
                "expire": {
                    "description": "过期时间：天",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "description": "443",
                    "type": "integer"
                },
                "server_id": {
                    "type": "integer"
                },
                "target": {
                    "$ref": "#/definitions/mmysql.ClientTarget"
                }
            }
        },
        "mparam.AddRelay": {
            "type": "object",
            "required": [
                "host",
                "name",
                "out_port",
                "port"
            ],
            "properties": {
                "host": {
                    "description": "api.github.com",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "out_port": {
                    "description": "443",
                    "type": "integer"
                },
                "port": {
                    "description": "443",
                    "type": "integer"
                }
            }
        },
        "mparam.AddResource": {
            "type": "object",
            "required": [
                "host",
                "name",
                "port"
            ],
            "properties": {
                "host": {
                    "description": "api.github.com,192.168.1.1/16",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "description": "80-443;3306;6379",
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "cidr",
                        "dns"
                    ]
                }
            }
        },
        "mparam.AddServer": {
            "type": "object",
            "required": [
                "host",
                "name",
                "out_port",
                "port"
            ],
            "properties": {
                "host": {
                    "description": "api.github.com",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "out_port": {
                    "description": "443",
                    "type": "integer"
                },
                "port": {
                    "description": "443",
                    "type": "integer"
                },
                "resource_id": {
                    "type": "string"
                }
            }
        },
        "mparam.EditClient": {
            "type": "object",
            "required": [
                "expire",
                "id",
                "name",
                "port",
                "server_id",
                "target"
            ],
            "properties": {
                "expire": {
                    "description": "过期时间：天",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "description": "443",
                    "type": "integer"
                },
                "server_id": {
                    "type": "integer"
                },
                "target": {
                    "$ref": "#/definitions/mmysql.ClientTarget"
                }
            }
        },
        "mparam.EditRelay": {
            "type": "object",
            "required": [
                "host",
                "id",
                "name",
                "out_port",
                "port"
            ],
            "properties": {
                "host": {
                    "description": "api.github.com",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "out_port": {
                    "description": "443",
                    "type": "integer"
                },
                "port": {
                    "description": "443",
                    "type": "integer"
                }
            }
        },
        "mparam.EditResource": {
            "type": "object",
            "required": [
                "host",
                "id",
                "name",
                "port"
            ],
            "properties": {
                "host": {
                    "description": "api.github.com,192.168.1.1/16",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "description": "80-443;3306;6379",
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "cidr",
                        "dns"
                    ]
                }
            }
        },
        "mparam.EditServer": {
            "type": "object",
            "required": [
                "host",
                "id",
                "name",
                "out_port",
                "port"
            ],
            "properties": {
                "host": {
                    "description": "api.github.com",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "out_port": {
                    "description": "443",
                    "type": "integer"
                },
                "port": {
                    "description": "443",
                    "type": "integer"
                },
                "resource_id": {
                    "type": "string"
                }
            }
        }
    }
}